Add auto-resize function when flashing system.img [...]
When flashing a system.img, SHRP will now enlarge the filesystem to fit
the full block device.

This is for flashing GSI's, which are built with a minimum /system size
but enlarge themselves on first boot to fit the device. Adding this logic
to SHRP makes it much easier for users to update a GSI ROM.

The auto-resize will be automatically performed, providing:
a) A sparse system.img is flashed;
b) The "Skip auto-resize" option is left unchecked (on the confirm flash screen); and
c) The "Mount system as read-only" option in mounts is NOT enabled.